requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6

SQLAlchemy==2.0.34
psycopg[binary]==3.2.1

pydantic==2.9.2
pydantic-settings==2.5.2

python-multipart==0.0.9

passlib[bcrypt]==1.7.4
PyJWT==2.9.0

email-validator==2.2.0
requests==2.32.3